R1: refactor(CA-904): share the search results loading/empty views via libraries/global_search
Move SearchResultsLoadingView and SearchResultsEmptyView from
features/global_search into libraries/global_search/presentation/widgets
(the established shared home, next to SearchLoadFailureWidget) and delete
the duplicated ProjectSearchResultsLoadingView/EmptyView so both
GlobalSearchPage and ProjectSearchPage render the same widgets. The
shared SearchResultsEmptyView already carries golden coverage from CA-899,
closing the empty-state golden gap flagged in review.
ProjectSearchResultsList stays feature-owned (domain-specific card).
