+++ lib/features/global_search/presentation/widgets/search_results_views.dart
@@ -6,8 +6,6 @@ import 'package:construculator/libraries/extensions/extensions.dart';
 import 'package:flutter/material.dart';
 import 'package:ripplearc_coreui/ripplearc_coreui.dart';
 
-const double _emptyStateMaxMessageWidth = 320.0;
-
 /// Scrollable list of search results grouped under a "Most relevant" header.
 ///
 /// Renders [EstimationCard] for each estimation in [results]. The
@@ -82,57 +80,6 @@ class SearchResultsList extends StatelessWidget {
           ),
         ),
       ],
-    );
-  }
-}
-
-/// Centered loading indicator shown while a search request is in flight.
-class SearchResultsLoadingView extends StatelessWidget {
-  const SearchResultsLoadingView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      key: Key('searchResultsLoadingView'),
-      child: CoreLoadingIndicator(key: Key('loadingIndicator')),
-    );
-  }
-}
-
-/// Empty state shown when a search completes with no matching results.
-class SearchResultsEmptyView extends StatelessWidget {
-  /// The query that produced no results; shown in the message.
-  final String query;
-
-  const SearchResultsEmptyView({super.key, required this.query});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      key: const Key('searchResultsEmptyView'),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          CoreIconWidget(
-            key: const Key('emptySearchIcon'),
-            icon: CoreIcons.fileSearch,
-            size: CoreIconSize.size32,
-            color: context.colorTheme.iconGrayMid,
-          ),
-          const SizedBox(height: CoreSpacing.space6),
-          ConstrainedBox(
-            constraints: const BoxConstraints(maxWidth: _emptyStateMaxMessageWidth),
-            child: Text(
-              context.l10n.searchResultsEmpty(query),
-              key: const Key('emptySearchMessage'),
-              textAlign: TextAlign.center,
-              style: context.textTheme.bodyMediumRegular.copyWith(
-                color: context.colorTheme.textHeadline,
-              ),
-            ),
-          ),
-        ],
-      ),
     );
   }
 }
