+++ test/features/global_search/units/accessibility/search_results_views_a11y_test.dart
@@ -3,6 +3,7 @@ import 'package:construculator/features/global_search/presentation/widgets/searc
 import 'package:construculator/l10n/generated/app_localizations.dart';
 import 'package:construculator/libraries/estimation/domain/entities/cost_estimate_entity.dart';
 import 'package:construculator/libraries/estimation/testing/testing.dart';
+import 'package:construculator/libraries/global_search/presentation/widgets/search_results_views.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
