+++ test/features/global_search/units/widgets/search_results_list_test.dart
@@ -4,6 +4,7 @@ import 'package:construculator/features/global_search/presentation/widgets/searc
 import 'package:construculator/l10n/generated/app_localizations.dart';
 import 'package:construculator/libraries/estimation/domain/entities/cost_estimate_entity.dart';
 import 'package:construculator/libraries/estimation/testing/testing.dart';
+import 'package:construculator/libraries/global_search/presentation/widgets/search_results_views.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:ripplearc_coreui/ripplearc_coreui.dart';
